Add dispatch for optimizers
