+++ Examples/CustomHamiltonian/custom_hamiltonian.py
@@ -43,7 +43,7 @@ ma.init_random_parameters(seed=1234, sigma=0.01)
 sa = nk.sampler.MetropolisLocal(machine=ma)
 
 # Optimizer
-op = nk.optimizer.AdaDelta()
+op = nk.optimizer.AdaDelta(ma)
 
 # Stochastic reconfiguration
 gs = nk.variational.Vmc(
