+++ Examples/Heisenberg1d/heisenberg1d_ffnn.py
@@ -41,7 +41,7 @@ sa = nk.sampler.MetropolisExchange(machine=ffnn)
 
 
 # Optimizer
-op = nk.optimizer.Sgd(learning_rate=0.05)
+op = nk.optimizer.Sgd(ma, learning_rate=0.05)
 
 # Stochastic reconfiguration
 gs = nk.variational.Vmc(
