+++ Examples/Heisenberg1d/heisenberg1d_jastrow.py
@@ -34,7 +34,7 @@ ma.init_random_parameters(seed=1234, sigma=0.01)
 sa = nk.sampler.MetropolisExchange(machine=ma)
 
 # Optimizer
-op = nk.optimizer.Sgd(learning_rate=0.05)
+op = nk.optimizer.Sgd(ma, learning_rate=0.05)
 
 # Stochastic reconfiguration
 gs = nk.Vmc(
