+++ Examples/Ising1d/ising1d_jastrow.py
@@ -31,7 +31,7 @@ ma.init_random_parameters(seed=1234, sigma=0.01)
 sa = nk.sampler.MetropolisLocal(machine=ma)
 
 # Optimizer
-op = nk.optimizer.Sgd(learning_rate=0.1)
+op = nk.optimizer.Sgd(ma, learning_rate=0.1)
 
 # Stochastic reconfiguration
 gs = nk.variational.Vmc(
