+++ netket/optimizer/ada_grad.py
@@ -1,8 +1,9 @@
-from .abstract_optimizer import AbstractOptimizer
-import numpy as _np
+from functools import singledispatch
+from . import numpy
 
 
-class AdaGrad(AbstractOptimizer):
+@singledispatch
+def AdaGrad(machine, learning_rate=0.001, epscut=1.0e-7):
     r"""AdaGrad Optimizer.
         In many cases, in Sgd the learning rate :math`\eta` should
         decay as a function of training iteration to prevent overshooting
@@ -19,49 +20,17 @@ class AdaGrad(AbstractOptimizer):
         AdaGrad has been shown to perform particularly well when
         the gradients are sparse, but the learning rate may become too small
         after many updates because the sum over the squares of past gradients is cumulative.
-    """
 
-    def __init__(self, learning_rate=0.001, epscut=1.0e-7):
-        r"""
-           Constructs a new ``AdaGrad`` optimizer.
 
-           Args:
-               learning_rate: Learning rate :math:`\eta`.
-               epscut: Small :math:`\epsilon` cutoff.
+        Args:
+           learning_rate: Learning rate :math:`\eta`.
+           epscut: Small :math:`\epsilon` cutoff.
 
-           Examples:
-               Simple AdaGrad optimizer.
+        Examples:
+           Simple AdaGrad optimizer.
 
-               >>> from netket.optimizer import AdaGrad
-               >>> op = AdaGrad()
+           >>> from netket.optimizer import AdaGrad
+           >>> op = AdaGrad()
         """
 
-        self._eta = learning_rate
-        self._epscut = epscut
-
-        if epscut <= 0:
-            raise ValueError("Invalid epsilon cutoff.")
-        if learning_rate < 0:
-            raise ValueError("Invalid learning rate.")
-
-        self._Gt = None
-
-    def update(self, grad, pars):
-        if self._Gt is None:
-            self._Gt = _np.zeros(pars.shape[0])
-
-        self._Gt += _np.abs(grad) ** 2
-
-        pars -= self._eta * grad / _np.sqrt(self._Gt + self._epscut)
-
-        return pars
-
-    def reset(self):
-        if self._mt is not None:
-            self._Gt.fill(0.0)
-
-    def __repr__(self):
-        rep = "AdaGrad optimizer with these parameters :"
-        rep += "\nLearning Rate = " + str(self._eta)
-        rep += "\nepscut = " + str(self._epscut)
-        return rep
+    return numpy.AdaGrad(learning_rate, epscut)
