+++ netket/optimizer/ada_max.py
@@ -1,8 +1,9 @@
-from .abstract_optimizer import AbstractOptimizer
-import numpy as _np
+from functools import singledispatch
+from . import numpy
 
 
-class AdaMax(AbstractOptimizer):
+@singledispatch
+def AdaMax(machine, alpha=0.001, beta1=0.9, beta2=0.999, epscut=1.0e-7):
     r"""AdaMax Optimizer.
 
         AdaMax is an adaptive stochastic gradient descent method,
@@ -20,72 +21,19 @@ class AdaMax(AbstractOptimizer):
         For an in-depth description of this method, please refer to
         `Kingma, D., & Ba, J. (2015). Adam: a method for stochastic optimization <https://arxiv.org/pdf/1412.6980.pdf>`_
         (Algorithm 2 therein).
-    """
 
-    def __init__(self, alpha=0.001, beta1=0.9, beta2=0.999, epscut=1.0e-7):
-        r"""
-           Constructs a new ``AdaMax`` optimizer.
 
-           Args:
-               alpha: The step size.
-               beta1: First exponential decay rate.
-               beta2: Second exponential decay rate.
-               epscut: Small epsilon cutoff.
+        Args:
+           alpha: The step size.
+           beta1: First exponential decay rate.
+           beta2: Second exponential decay rate.
+           epscut: Small epsilon cutoff.
 
-           Examples:
-               Simple AdaMax optimizer.
+        Examples:
+           Simple AdaMax optimizer.
 
-               >>> from netket.optimizer import AdaMax
-               >>> op = AdaMax()
+           >>> from netket.optimizer import AdaMax
+           >>> op = AdaMax()
         """
 
-        if epscut <= 0:
-            raise ValueError("Invalid epsilon cutoff.")
-        if alpha < 0:
-            raise ValueError("Invalid alpha.")
-        if beta1 < 0 or beta1 > 1:
-            raise ValueError("Invalid beta1.")
-        if beta2 < 0 or beta2 > 1:
-            raise ValueError("Invalid beta1.")
-
-        self._alpha = alpha
-        self._beta1 = beta1
-        self._beta2 = beta2
-        self._epscut = epscut
-
-        self._mt = None
-        self._ut = None
-        self._niter = 0
-
-    def update(self, grad, pars):
-        if self._mt is None:
-            self._mt = _np.zeros(pars.shape[0])
-            self._ut = _np.zeros(pars.shape[0])
-
-        self._mt = self._beta1 * self._mt + (1.0 - self._beta1) * grad
-
-        self._ut = _np.maximum(
-            _np.maximum(_np.abs(grad), self._beta2 * self._ut), self._epscut
-        )
-
-        self._niter += 1
-
-        eta = self._alpha / (1.0 - self._beta1 ** self._niter)
-
-        pars -= eta * self._mt / self._ut
-
-        return pars
-
-    def reset(self):
-        if self._mt is not None:
-            self._mt.fill(0.0)
-            self._ut.fill(0.0)
-            self._niter = 0
-
-    def __repr__(self):
-        rep = "AdaMax optimizer with these parameters :"
-        rep += "\nStep size alpha = " + str(self._alpha)
-        rep += "\nbeta1 = " + str(self._beta1)
-        rep += "\nbeta2 = " + str(self._beta2)
-        rep += "\nepscut = " + str(self._epscut)
-        return rep
+    return numpy.AdaMax(alpha, beta1, beta2, epscut)
