+++ netket/optimizer/numpy/__init__.py
@@ -1,0 +1,9 @@
+from .ada_delta import AdaDelta
+from .ada_grad import AdaGrad
+from .ada_max import AdaMax
+from .ams_grad import AmsGrad
+from .momentum import Momentum
+from .rms_prop import RmsProp
+from .sgd import Sgd
+
+from .stochastic_reconfiguration import SR
